pyproject.toml
[project]
name = "icicle-playgrounds"
version = "0.1.3"
description = "SDK to interact with the ICICLE Model & Data Playgrounds"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "torchvision>=0.22.1",
]

[dependency-groups]
dev = [
    "docstring-to-markdown>=0.17",
    "hypothesis>=6.135.16",
    "pytest>=8.4.1",
]